requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
attrs==25.4.0
audioop-lts==0.2.2
colorama==0.4.6
discord-protos==0.0.2
discord.py==2.6.4
discord.py-self==2.0.1
frozenlist==1.8.0
idna==3.11
multidict==6.7.0
propcache==0.4.1
protobuf==6.33.2
yarl==1.22.0
